Clan leader can delete posts from his clan thread
Part 1

+++ Zero-K.info/Controllers/ForumController.cs
@@ -27,6 +27,20 @@ namespace ZeroKWeb.Controllers
         }
 
         /// <summary>
+        ///     Returns true if the thread is a clanthread and the user is the leader of this clan
+        /// </summary>
+        bool IsClanLeaderForThread(ForumThread thread) {
+            ZkDataContext db = new ZkDataContext();
+
+            Clan clan = db.Clans.FirstOrDefault(x => x.ForumThreadID == thread.ForumThreadID);
+            if (clan == null || Global.ClanID != clan.ClanID) return false;
+
+            if (!Global.Account.HasClanRight(x => x.RightKickPeople)) return false; // RightKickPeople is clan leader; could use a better name
+
+            return true;
+        }
+
+        /// <summary>
         ///     Set the last post time of the thread to current time (includes edits)
         /// </summary>
         void ResetThreadLastPostTime(int threadID) {
@@ -54,7 +68,7 @@ namespace ZeroKWeb.Controllers
         }
 
         [HttpPost]
-        [Auth(Role = AdminLevel.Moderator)]
+        [Auth]
         [ValidateAntiForgeryToken]
         public ActionResult DeletePost(int? postID, bool deleteThread = false) {
             var db = new ZkDataContext();
@@ -65,6 +79,8 @@ namespace ZeroKWeb.Controllers
             var threadID = thread.ForumThreadID;
             //int index = post.ForumThread.ForumPosts.IndexOf(post);
             var page = GetPostPage(post);
+
+            if (!Global.IsModerator && !IsClanLeaderForThread(thread)) return Content("No permission to delete this post");
 
             db.ForumPosts.DeleteOnSubmit(post);
             if ((thread.ForumPosts.Count() <= 0 || deleteThread) && IsNormalThread(thread))
